// src/types/AirFareLineItemType.ts
type AirFareLineItemType = {
  jobId: string;
  SearchName: string;
  vertical: string;
  sMarkets: string;
  specap1: string;
  specap2: string;
  specapT1: string;
  specapT2: string;
  sites: string;
  spcxr: string;
  ref: string;
  flc: string;
  pas: string;
  wcn: string;
  bke: string;
  sDateRangeStart: string;
  sDateRangeEnd: string;
  sDOWDepart: string;
  sDOWReturn: string;
  sDirectOnly: string;
  sDepartTime: string;
  sReturnTime: string;
  sSearchPOS: string;
  sSpecificStops: string;
  custom: string;
  geo: string;
  dateRangeLength: string;
  dateRangeFixedReturn: string;
  sSearchType: string;
  ap1: boolean;
  ap7: boolean;
  ap14: boolean;
  ap21: boolean;
  ap28: boolean;
  ap1s: boolean;
  ap7s: boolean;
  ap14s: boolean;
  ap21s: boolean;
  ap28s: boolean;
  sIncCodeshares: boolean;
  searchName?: string;
  dateRange?: string;
  pointOfSale?: string;
  searchType?: string;
  locationPickup?: string;
  locationDropoff?: string;
  pickTime?: string;
  dropTime?: string;
  limitOfMonthPickup?: string;
  limitOfMonthReturn?: string;
  limitOfWeekPickup?: string;
  limitOfWeekReturn?: string;
  discount?: string;
  country?: string;
  address?: string;
  city?: string;
  state?: string;
  postalCode?: string;
  searchRadius?: string | number;
  RAC?: string;
  locations?: any;
};

export default AirFareLineItemType;